Moved RAM class defination to it's own file
